+++ src/test/converter2/behavior/mergedDeclarations.ts
@@ -1,0 +1,8 @@
+export namespace SingleCommentMultiDeclaration {}
+/** Comment on second declaration */
+export namespace SingleCommentMultiDeclaration {}
+
+/** Comment 1 */
+export namespace MultiCommentMultiDeclaration {}
+/** Comment 2 */
+export namespace MultiCommentMultiDeclaration {}